Refactoring tests to run in parallel
