+++ circle.yml
@@ -1,3 +1,8 @@
+# Set up the machine
+machine:
+    environment:
+        ZPAR_MODEL_DIR: ${HOME}/english-models
+
 # Install stuff
 dependencies:
   cache_directories:
@@ -12,13 +17,13 @@ dependencies:
     - ${HOME}/miniconda3/bin/conda config --add channels desilinguist
     - ${HOME}/miniconda3/bin/conda update --yes conda
     - ${HOME}/miniconda3/bin/conda install --yes six nose nltk
-    - ${HOME}/miniconda3/bin/python -m nltk.downloader wordnet -d ${HOME}/nltk_data
+    - mkdir -p ${HOME}/nltk/data
+    - ${HOME}/miniconda3/bin/python -m nltk.downloader wordnet -d ${HOME}/nltk/data
     - make python-zpar
     - ${HOME}/miniconda3/bin/pip install -e .
 
 # Run test
 test:
   override:
-    - ZPAR_MODEL_DIR=${HOME}/english-models ${HOME}/miniconda3/bin/nosetests -v tests/test_tagger.py tests/test_parser.py tests/test_depparser.py
-    - rm -rf ${HOME}/nltk_data
-    - ZPAR_MODEL_DIR=${HOME}/english-models ${HOME}/miniconda3/bin/nosetests -v tests/test_depparser_no_wordnet.py
+    - case $CIRCLE_NODE_INDEX in 0) ${HOME}/miniconda3/bin/nosetests -v tests/test_tagger.py ;; 1) ${HOME}/miniconda3/bin/nosetests -v tests/test_parser.py ;; 2) ${HOME}/miniconda3/bin/nosetests -v tests/test_depparser.py ;; 3) NLTK_DATA= ${HOME}/miniconda3/bin/nosetests -v tests/test_depparser_no_wordnet.py ;; esac:
+        parallel: true
